Add pseudo-CAD mode for end nodes
CAD (channel activity detection) allows a node to monitor a channel for
incoming transmissions and automatically switch to the appropriate SF to
receive a packet. While this functionality is not implemented in the flora
framework, this patch emulates the CAD behaviour on a node, including a
configurable sensitivity reduction.

Additionally, change a bunch of sequential "if"s into nested "else if"s.

Signed-off-by: Roger Pueyo Centelles <rpueyo@ac.upc.edu>